the gamma patch
I'll push the pkgbuild that uses this once this specific
instance has been tested to work. Anyways it should work...
ish. The hardware makes it (almost?) impossible to switch
from one gamma table to another without a frame without gamma
applied in between, and i've tried 2 things already, and
like... i dunno if things 3 and 4 would be worth the energy.